refactor init
