+++ src/api/helpers/setColumnsValue.js
@@ -1,0 +1,14 @@
+/**
+ * 
+ * @param {*} columns : Table columns ;
+ * @returns Default Time and Value column headers
+ */
+export function setColumnsTsValue(columns) {
+    if (columns.length > 0) {
+        return [
+            { Header: "Time", accessor: "time" },
+            ...columns,
+            { Header: "Value", accessor: "value" },
+        ];
+    } else return [];
+}
