+++ src/components/VectorTable/helpers/index.js
@@ -1,0 +1,5 @@
+export * from "./scrollbarWidth";
+export * from "./prepareCols";
+export * from "./prepareVectorRows";
+export * from "./setColumnsData";
+export * from "./setColumnsValue";
